Move GameResult class to libs/domain/score.py and update import paths for consistency
