+++ integrations/slack/events/comparison.py
@@ -6,12 +6,12 @@ import logging
 from typing import TYPE_CHECKING, cast
 
 import libs.global_value as g
-from cls.score import GameResult
 from cls.timekit import ExtendedDatetime as ExtDt
 from cls.timekit import Format
 from integrations.protocols import ActionStatus, CommandType
 from libs.data import lookup, modify, search
 from libs.datamodels import ComparisonResults
+from libs.domain.score import GameResult
 from libs.types import StyleOptions
 from libs.utils import formatter, validator
