+++ libs/dispatcher.py
@@ -7,10 +7,10 @@ import re
 from typing import TYPE_CHECKING
 
 import libs.global_value as g
-from cls.score import GameResult
 from integrations import factory
 from integrations.protocols import MessageStatus
 from libs.data import lookup, modify
+from libs.domain.score import GameResult
 from libs.functions import message
 from libs.types import StyleOptions
 from libs.utils import formatter, validator
